refactor: better folder structure for chat
