+++ src/routes/chat/Chat.jsx
@@ -1,47 +1,14 @@
 import React from 'react';
-import { useParams, Link, Redirect } from 'react-router-dom';
+import { useParams, Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 
 import PageContainer from 'components/pageContainer';
 import { CHAT } from 'constants/routes';
 
-import ChatRooms from './ChatRooms';
-import ChatMessages from './ChatConversation/ChatMessages';
-import ChatInput from './ChatInput/ChatInput';
-import ChatHeader from './ChatHeader';
+import ChatNav from './ChatNav';
+import ChatList from './chatList';
+import ChatContent from './chatContent';
 import './Chat.scss';
-
-const ChatNav = ({ isShowingChatRoom, headerText }) => {
-  return (
-    <div className="chat__header columns">
-      <div className="column is-full-mobile is-two-fifths-tablet chat__header__left">
-        {isShowingChatRoom && (
-          <Link
-            to={CHAT}
-            className="chat__header__back button button--cta button--nav--circle"
-            type="button"
-          >
-            <i className="fas fa-arrow-left" />
-          </Link>
-        )}
-        <span>Matches</span>
-      </div>
-      <div className="column is-hidden-mobile chat__header__right">
-        <span className="chat__header__right--text">{headerText}</span>
-      </div>
-    </div>
-  );
-};
-
-const ChatContent = ({ chat }) => {
-  return (
-    <div className="column chat__content">
-      <ChatHeader chat={chat} />
-      <ChatMessages chat={chat} />
-      <ChatInput chat={chat} />
-    </div>
-  );
-};
 
 const Chat = () => {
   const { chatRoomId } = useParams();
@@ -59,7 +26,7 @@ const Chat = () => {
         headerText={chatNavHeaderText}
       />
       <div className="columns is-mobile is-gapless">
-        <ChatRooms isShowingChatRoom={!!chatRoomId} />
+        <ChatList isShowingChatRoom={!!chatRoomId} />
         {chatRoomId && <ChatContent chat={chat} />}
       </div>
     </PageContainer>
