--- src/routes/chat/ChatRooms/index.js
@@ -1,1 +1,0 @@
-export { default } from './ChatRooms';
