+++ src/routes/chat/chatContent/chatHeader/ChatOfferSubheader.scss
