+++ src/routes/chat/chatContent/chatInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatInput';
