+++ src/routes/chat/chatContent/chatMessages/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatMessagesContainer';
