+++ src/routes/chat/chatContent/chatMessages/message/chat/ChatMessage.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+import { getTimeFromTimestamp } from 'utils';
+
+import './ChatMessage.scss';
+
+const ChatMessage = ({ message }) => {
+  const timeString = getTimeFromTimestamp(message.createdAt);
+
+  return (
+    <p className="chatMessage__bubble__message">
+      <span className="chatMessage__bubble__message--message">
+        {message.message}
+      </span>
+      <span className="chatMessage__bubble__message--timestamp">
+        {timeString}
+      </span>
+    </p>
+  );
+};
+
+export default ChatMessage;
