+++ src/routes/chat/chatContent/chatMessages/message/chat/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatMessage';
