+++ src/routes/chat/chatContent/chatMessages/message/index.js
@@ -1,0 +1,1 @@
+export { default } from './Message';
