+++ src/routes/chat/chatContent/chatMessages/message/offer/OfferMessage.scss
