+++ src/routes/chat/chatContent/chatMessages/message/offer/index.js
@@ -1,0 +1,1 @@
+export { default } from './OfferMessage';
