+++ src/routes/chat/chatContent/successfulMatch/index.js
@@ -1,0 +1,1 @@
+export { default } from './SuccessfulMatchContainer';
