+++ src/routes/chat/chatList/ChatList.jsx
@@ -5,11 +5,11 @@ import { useSelector } from 'react-redux';
 
 import EmptyMessagesSvg from 'components/svgr/EmptyMessages';
 
-import ChatRoom from './ChatRoom';
-import './ChatRooms.scss';
+import ChatListItem from './ChatListItem';
+import './ChatList.scss';
 
-const ChatRooms = ({ isShowingChatRoom }) => {
-  const chatRooms = useSelector(state => {
+const ChatList = ({ isShowingChatRoom }) => {
+  const chatList = useSelector(state => {
     return Object.values(state.chat.unarchived).sort(
       (a, b) => b.updatedAt - a.updatedAt
     );
@@ -18,7 +18,7 @@ const ChatRooms = ({ isShowingChatRoom }) => {
   const { url } = useRouteMatch();
   const basePath = getCurrentPathWithoutParam(url);
 
-  if (chatRooms.length === 0) {
+  if (chatList.length === 0) {
     return (
       <ul
         className={`chatrooms column ${
@@ -39,11 +39,11 @@ const ChatRooms = ({ isShowingChatRoom }) => {
         isShowingChatRoom ? 'is-hidden-mobile' : 'is-full-mobile'
       } is-two-fifths-tablet`}
     >
-      {chatRooms.map(chat => (
-        <ChatRoom key={chat.id} chat={chat} basePath={basePath} />
+      {chatList.map(chat => (
+        <ChatListItem key={chat.id} chat={chat} basePath={basePath} />
       ))}
     </ul>
   );
 };
 
-export default ChatRooms;
+export default ChatList;
