+++ src/routes/chat/chatList/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChatList';
